Add feature to model assortativity
Assortativity affects the way individuals interact to play the game. It
is the tendency for an agent to preferentially play with individuals
who have the same strategy. Assortativity in this program is controlled
with parameter $\alpha \in $ [0,1]. If assortativity $\alpha $ is 0,
players interact at random, so the probability of meeting a player
using strategy $i$ equals the frequency of that strategy in the
population, i.e. $x_i$. At the other extreme, if assortativity $\alpha$
is 1, players interact only with players using the same strategy as
theirs, so a player using strategy $i$ will only meet players using
strategy $i$. For intermediate cases, $\alpha $ $\in$ (0,1), a player
will interact with a random player from the population with probability
$(1-\alpha)$, and with a player with the same strategy as his with
probability $\alpha $.